requirements.txt
einops
numpy==1.24.3
torch
torchvision
pillow
transformers
accelerate>=0.26.0
bitsandbytes
typing-extensions==4.8.0
flash-attn --no-build-isolation
